Fix selecting underexposed images multiple times
