+++ app/src/main/java/com/uf/biolens/ui/imageGrid/ImageGridActivity.kt
@@ -39,7 +39,7 @@ import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.uf.biolens.R
 import com.uf.biolens.data.BioLensRepository
 import com.uf.biolens.databinding.ActivityImageGridBinding
-import com.uf.biolens.imaging.getUnderexposedImages
+import com.uf.biolens.imaging.UnderexposedImageFinder
 import com.uf.biolens.network.GoogleDriveUploadWorker
 import com.uf.biolens.network.GoogleSignInHelper
 import com.uf.biolens.ui.common.ExportOptions
@@ -401,11 +401,16 @@ class ImageGridActivity : AppCompatActivity(), ImageSelectorListener {
         if (selectUnderexposedJob != null) {
             return
         }
+        val finder = UnderexposedImageFinder(sessionID)
+        finder.underexposedImageHandler = { image ->
+            viewModel.imageSelector.setSelected(image, true)
+            adapter.refreshEditingState()
+        }
+        finder.onCompleteListener = {
+            selectUnderexposedJob = null
+        }
         selectUnderexposedJob = lifecycleScope.launch {
-            getUnderexposedImages(sessionID) { image ->
-                viewModel.imageSelector.setSelected(image, true)
-                adapter.refreshEditingState()
-            }
+            finder.getUnderexposedImages()
         }
     }
